fix addMovieToLocalStorage

+++ modules/storage.js
@@ -1,5 +1,8 @@
 export function addMovieToLocalStorage(key, value) {
   const data = JSON.parse(localStorage.getItem(key)) || [];
-  data.push(value);
-  localStorage.setItem(key, JSON.stringify(data));
+  const existingMovie = data.find(movie => movie.id === value.id)
+  if (!existingMovie) {
+    data.push(value);
+      localStorage.setItem(key, JSON.stringify(data));
+  } 
 }
